requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
prophet==1.1.5
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
